Bugs about the one-dimensional wet-dry interface assemble, and the same correction has been made for the two-dimensional non-hydrostatic solver

+++ NdgPhys/NdgMatSolver/NdgNonhydrostaticSolver/@NdgNonhydrostaticSolver1d/matAssembleWetDryInterface1d.m
@@ -4,7 +4,7 @@ function matAssembleWetDryInterface1d(obj, mesh)
 %> Function to assemble the wet and dry interface
 %> @param[in] mesh The mesh object
 
-[ obj.WetDryCell ] = mxAssembleWetDryInterface1d(mesh.status, mesh.EToE, ...
+[ obj.WetDryCell ] = mxAssembleWetDryInterface1d(int8(mesh.status), mesh.EToE, ...
     mesh.InnerEdge.FToE, mesh.BoundaryEdge.FToE, mesh.BoundaryEdge.ftype, mesh.BoundaryEdge.FToF);
